Separate class for config dump handler
